pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow==11.0.0",
    "tkcalendar>=1.6.1",
]
